Fix warning when building with gcc 4.8.3
Cast to int the value given to isspace(3) so that we suppress the
warning below and fix the build on NetBSD 6.99.31 with gcc 4.8.3.

tap_parser.c: In function 'kyua_tap_try_parse_plan':
tap_parser.c:253:9: error: array subscript has type 'char' [-Werror=char-subscripts]
         while (*reason != '\0' && isspace(*reason))

Not completely sure it's the right fix though.

+++ kyua-testers/tap_parser.c
@@ -250,7 +250,7 @@ kyua_tap_try_parse_plan(const char* line, kyua_tap_summary_t* summary)
         skip_start = strstr(line, "skip");
     if (skip_start != NULL) {
         const char *reason = skip_start + strlen("SKIP");
-        while (*reason != '\0' && isspace(*reason))
+        while (*reason != '\0' && isspace((int)*reason))
             ++reason;
         if (*reason == '\0') {
             summary->all_skipped_reason = strdup("No reason specified");
